Add nethsm list-users subcommand
This patch adds the nethsm list-users subcommand that lists all users on
the NetHSM device.  Per default, it lists the user ID, the real name and
the role (requiring one API request per user), but it can also be
restricted to only print the user ID (with only one API request).