nk3: Fix reboot subcommand for macOS
After we send a reboot command to the device, we lose the connection
while it is rebooting.  This causes an error which we want to ignore
because it is expected behavior.  On Linux, the raised error is an
OSError with errno EIO = 5, so we currently check for that.  But on
macOS, an OSError without errno is raised, so we remove the errno check
and ignore all OSError instances.  To make debugging easier, we also add
a log message printing the ignored exception.