Add type stubs for fido2
This patch adds stubs for the fido2 library that are used by mypy for
the static type analysis.  This is necessary as fido2 does not have type
annotations and is not part of the typeshed repository.  Our type stubs
only contain definitions for the functions that are used in typed code
and are by no means complete.