fido2: Fix serial number selection
The fido2 library added the serial number to the device descriptor in
version 0.9.3, so we have to bump this dependency for the --serial
option to work reliably.  This patch also fixes the find_device method
to properly query the serial number.  (The descriptor is a named tuple,
not a dictionary.)

Fixes #101.