cli: Add CliException for easier error handling
Currently, the local_critical function defined in pynitrokey.helpers is
the preferred way to report a critical error.  Per default, it prints a
support message mentioning the log file location and terminates the
program.  The problem with this approach is that it is not clear from
the API that the function call terminates the program.  This affects
static analysis and code readability.

This patch introduces the CliException, a subclass of
click.ClickException.  Raising this exception from a click command has
the same effect as calling the local_critical function while making it
clear that the regular program flow is disrupted.

Fixes https://github.com/Nitrokey/pynitrokey/issues/135