nk3 update: Handle bootloader connection issues
Users report issues with the bootloader connection after the reboot.
Apparently, the udev rules are not yet applied when we try to connect to
the device.  We mitigate this issue by performing up to three retries.

This also adds a reference to the udev rules to the error message if
nitropy is running on Linux.

Fixes https://github.com/Nitrokey/pynitrokey/issues/167