Remove duplicated dependency list
This patch removes the duplicated dependency list in ci-requirements.txt
so that we now only use the dependencies listed in pyproject.toml.  They
can be installed with `flit -m install`.

This patch also removes the dev dependencies from the wine build script
because they are not required for the build process.