Rename ls command to list
For all devices, we are currently using list as the name for the
subcommand that lists the devices.  Only on the top level, we use ls
instead.  This patch makes the CLI more consistent by renaming the
top-level ls command to list and adding an alias with a deprecation
warning for the old name.