Fix error message for CliException
Previously, the CliException constructor assumed that all arguments can
be joined into a single string message.  This did not take into account
that we can also provied other objects, e. g. exceptions.  With this
patch, we stringify all arguments before joining them to avoid this
issue.

See [0] for an example of this issue.

[0] https://github.com/Nitrokey/pynitrokey/issues/192