Migrate mypy configuration into pyproject.toml
Since mypy 9.000, the mypy configuration can be specified in
pyproject.toml.  This format is much more legible as modules can be
specified as an array instead of concatenating them into one line.
There are also other improvements in the newer mypy versions.
Therefore, this patch bumps the minimum mypy version and moves the mypy
configuration into pyproject.toml.  This does not affect the end user as
mypy is development dependency.