Redirect prompts to stderr
This patch changes prompts from click.confirm, click.prompt or
pynitrokey.helpers.AskUser to be printed to stderr instead of stdout.
This makes sure that the user will see the prompt even if the output is
redirected to a file or piped to another process.  This patch does not
fix this issue for options with prompt=True as that is more complex.

Fixes #228