NetHSM: Rename reset to factory-reset
This patch renames the reset command to factory-reset and updates it for
the new API that also changed the name of the command.

It also updates the API definition.  As upstream has added the required
produces tags to their API definition, we can directly download it and
no longer need to modify it manually.