Update documentation for supported Python versions
Previously, we decided to only support the Python versions present in
the latest Ubuntu LTS and regular releases or in the latest Debian
stable release.  We were not aware that upgrading Ubuntu LTS
installations is only recommended after there has been a point release
for the next LTS release.  We address this in this patch by updating the
rules for supported Python versions in the developer guide.