Relax black version requirement
Previously, we pinned the black version to 21.12b0 because newer
versions required click 8 which was not supported by spsdk 1.6.3.  But
since spsdk 1.7.0 also works wich click 8, we can now relax this
requirement.

Note that not all black versions work with all click versions.  But this
is an issue with black’s dependency specifications that we cannot fix in
our package.  And as black is only a development dependency, that should
not affect users.