Drop cbor dependency
As discussed in [0] and [1], the cbor dependency is unmaintained and
should be replaced.  But it is actually only used in one command that is
not even active, and fido2.cbor provides a simple cbor implementation
that is sufficient for our needs.  So with this patch, we change the
(unused) fido2 probe subcommand to use fido2.cbor instead of cbor and
drop the cbor dependency.

Fixes: https://github.com/Nitrokey/pynitrokey/issues/93

[0] https://github.com/Nitrokey/pynitrokey/issues/93
[1] https://github.com/Nitrokey/pynitrokey/commit/bc45e74d377f28853ff05cfe926e5ac9b7b36267