Make pyscard optional
3b8538f601b1ac7def4d6cbc658b47a91e2c8f6b introduced the pyscard
dependency for implementing the firmware mode test.  Yet this dependency
has some additional build dependencies and no binary wheel for Linux.
As the firmware mode test is mostly relevant for production and not for
end users, we make the pyscard dependency and the firmware mode test
otional and put it behind the pcsc extra.

Fixes https://github.com/Nitrokey/pynitrokey/issues/273