nk3: Support provisioner app
This patch adds a hidden subcommand for the provisioner app to the nk3
command.  Currently, it only supports provisioning the FIDO2 attestation
certificate and key.  Besides importing the key and certificate, it also
performs some validation steps to ensure that the key and certificate
match and that the certificate meets the specified requirements.