nk3: Remove pre from Version.__hash__
If __eq__ returns True for two objects, their hash must be the same.
Version instances can be equal even if their pre field is different if
at least one of the versions is incomplete.  Therefore the current hash
implementation for Version is incorrect.

This patch simplifies the hash function by only considering the major,
minor and patch components.  This leads to more hash collissions, but
this is not a concern for us.

This patch also adds simple doctests to __hash__ based on the doctests
for __eq__.