nk3 test: Fix provisioner test
This patch adds the provisioner test to the default excludes for the
nitropy nk3 test command.  The reason for this is that it the test is
not relevant outside of production and is likely to fail if
gnupg/scdaemon is running or if there is an issue with pcscd.

It also changes the test setup so that the provisioner test case is
always listed, even if pyscard is not installed.  In this case, it is
marked as skipped.

Fixes https://github.com/Nitrokey/pynitrokey/issues/323