nk3 set-config: Add warning and confirmation prompt
With the upcoming firmware release for the Nitrokey 3, changing some
configuration values can have side effects, namely a reset of the opcard
application.  We will add new commands with additional information and
warnings in the future.  For the time being, this patch adds a warning
to the set-config command, indicating that it may cause data loss and
should only be used for testing and development.