Restrict hidapi dependency
hidapi release v0.14.0-post.4 does not work with our udev rules as it
switches back to the libusb backend instead of using hidraw.  Until we
have rolled out new udev rules or hidapi has migrated to hidraw for
good, this patch restricts hidapi to the versions using the hidraw
backend.

Fixes: https://github.com/Nitrokey/pynitrokey/issues/601