Only restrict hidapi dependency on Linux
We restrict the hidapi dependency to some specific versions to make sure
that the hidraw backend is used, see:
https://github.com/Nitrokey/pynitrokey/pull/602

But the hidapi versions that are now eligible are not providing wheels
for Python 3.13 so that they have to be compiled from source.  This is
an unnecessary complication for non-Linux platforms.  Therefore this
patch only applies the restriction on Linux.  Other platforms can use
any 0.14 release.