nethsm: Read host from NETHSM_HOST if not set
This patch changes all nethsm commands to read the host from the
NETHSM_HOST environment variable if it is not set on the command line.
This has two benefits:  It makes it easier to run multiple commands for
the same host.  And it fixes the problem that the required --host option
on the nethsm command group would make it impossible to run --help for
a subcommand without setting --host.

Example output if neither is set:

    $ nitropy nethsm lock
    Command line tool to interact with Nitrokey devices 0.7.3
    Critical error:
    Missing NetHSM host: set the --host option or the NETHSM_HOST environment variable