nk3 test: Support credential creation without PIN
Previously, we always required a PIN to create a FIDO2 credential if a
PIN has been set.  In firmware version v1.8.2-rc.1, we added support for
the makeCredUvNotRqd option which means that credentials can be created
without PIN (user verification).  This patch updates the fido2 test in
nk3 test to not require the PIN if the device supports makeCredUvNotRqd.

Fixes: https://github.com/Nitrokey/pynitrokey/issues/678