nkpk: Add admin-app commands
This patch moves the nk3 commands using admin-app features to the
trussed module so that they can also be used with the nkpk.
factory-reset-app is the only command that is disabled for the nkpk as
fido-authenticator is the only app and it does not support factory
reset.

I also took the chance to remove the --experimental flag for good for
these commands as they have been stabilized for a long time.